- removed module Html helper class dependency + HtmlTrait
